+++ backend/problems/src/ma1/arithmetics.rs
@@ -1,2 +1,3 @@
 mod decimal_numbers;
 mod negative_numbers;
+mod order_of_operations;