+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - You can now search for `public` or `private` in the web editor
 - The "publish every X" button only shows when there is something to publish
+- Added four `order_of_operations` problems
 
 ## [0.12.0] - 2026-07-25